Poll to re-mute mic on old devices
In Android API 28 and above, a broadcast occurs when the microphone
is unmuted. Since there is no such broadcast available on lower
API versions, poll the microphone state periodically (with randomness)
such that if any other app unmuted the microphone, we can mute it
again (despite not receiving a broadcast).