Prevent runtime crash by creating notification channel
The Google Play Store requires us to target API 29 (instead of our preferred API 28)
in order to upload an APK. However, merely targeting API 29 in the Android Manfiest XML
(and not changing any Java code) causes changes to how notifications are handled in
the underlying notification system at runtime, and the app (which ran fine when targeting
API 28) will fatally crash at launch (citing that the notification was created
with out a valid notification channel). Thank you, Google. Thank you, Android.

So, we can create a notification channel and assign the notification to it,
but we must also now build the app with platform API 26 or above (since API 25
does not have the notification channel class, and therefore wouldn't compile).